+++ src/videoipath_automation_tool/apps/inventory/app/get_device.py
@@ -1180,6 +1180,21 @@ class InventoryGetDeviceMixin:
         label: Optional[str] = None,
         device_id: Optional[str] = None,
         address: Optional[str] = None,
+        custom_settings_type: Optional[Literal["com.nevion.lvb440-1.0.0"]] = None,
+        config_only: bool = False,
+        label_search_mode: Literal[
+            "canonical_label", "factory_label_only", "user_defined_label_only"
+        ] = "canonical_label",
+        status_fetch_retry: int = STATUS_FETCH_RETRY_DEFAULT,
+        status_fetch_delay: int = STATUS_FETCH_DELAY_DEFAULT,
+    ) -> InventoryDevice[CustomSettings_com_nevion_lvb440_1_0_0]: ...
+
+    @overload
+    def get_device(
+        self,
+        label: Optional[str] = None,
+        device_id: Optional[str] = None,
+        address: Optional[str] = None,
         custom_settings_type: Optional[Literal["com.nevion.maxiva-0.1.0"]] = None,
         config_only: bool = False,
         label_search_mode: Literal[
@@ -1285,6 +1300,21 @@ class InventoryGetDeviceMixin:
         label: Optional[str] = None,
         device_id: Optional[str] = None,
         address: Optional[str] = None,
+        custom_settings_type: Optional[Literal["com.nevion.mock_cloud-0.1.0"]] = None,
+        config_only: bool = False,
+        label_search_mode: Literal[
+            "canonical_label", "factory_label_only", "user_defined_label_only"
+        ] = "canonical_label",
+        status_fetch_retry: int = STATUS_FETCH_RETRY_DEFAULT,
+        status_fetch_delay: int = STATUS_FETCH_DELAY_DEFAULT,
+    ) -> InventoryDevice[CustomSettings_com_nevion_mock_cloud_0_1_0]: ...
+
+    @overload
+    def get_device(
+        self,
+        label: Optional[str] = None,
+        device_id: Optional[str] = None,
+        address: Optional[str] = None,
         custom_settings_type: Optional[Literal["com.nevion.montone42-0.1.0"]] = None,
         config_only: bool = False,
         label_search_mode: Literal[
@@ -1323,6 +1353,21 @@ class InventoryGetDeviceMixin:
         status_fetch_retry: int = STATUS_FETCH_RETRY_DEFAULT,
         status_fetch_delay: int = STATUS_FETCH_DELAY_DEFAULT,
     ) -> InventoryDevice[CustomSettings_com_nevion_mwedge_0_1_0]: ...
+
+    @overload
+    def get_device(
+        self,
+        label: Optional[str] = None,
+        device_id: Optional[str] = None,
+        address: Optional[str] = None,
+        custom_settings_type: Optional[Literal["com.nevion.ndi-0.1.0"]] = None,
+        config_only: bool = False,
+        label_search_mode: Literal[
+            "canonical_label", "factory_label_only", "user_defined_label_only"
+        ] = "canonical_label",
+        status_fetch_retry: int = STATUS_FETCH_RETRY_DEFAULT,
+        status_fetch_delay: int = STATUS_FETCH_DELAY_DEFAULT,
+    ) -> InventoryDevice[CustomSettings_com_nevion_ndi_0_1_0]: ...
 
     @overload
     def get_device(
@@ -1465,6 +1510,21 @@ class InventoryGetDeviceMixin:
         label: Optional[str] = None,
         device_id: Optional[str] = None,
         address: Optional[str] = None,
+        custom_settings_type: Optional[Literal["com.nevion.nxl_me80-1.0.0"]] = None,
+        config_only: bool = False,
+        label_search_mode: Literal[
+            "canonical_label", "factory_label_only", "user_defined_label_only"
+        ] = "canonical_label",
+        status_fetch_retry: int = STATUS_FETCH_RETRY_DEFAULT,
+        status_fetch_delay: int = STATUS_FETCH_DELAY_DEFAULT,
+    ) -> InventoryDevice[CustomSettings_com_nevion_nxl_me80_1_0_0]: ...
+
+    @overload
+    def get_device(
+        self,
+        label: Optional[str] = None,
+        device_id: Optional[str] = None,
+        address: Optional[str] = None,
         custom_settings_type: Optional[Literal["com.nevion.openflow-0.0.1"]] = None,
         config_only: bool = False,
         label_search_mode: Literal[
@@ -1600,6 +1660,21 @@ class InventoryGetDeviceMixin:
         label: Optional[str] = None,
         device_id: Optional[str] = None,
         address: Optional[str] = None,
+        custom_settings_type: Optional[Literal["com.nevion.spg9000-0.1.0"]] = None,
+        config_only: bool = False,
+        label_search_mode: Literal[
+            "canonical_label", "factory_label_only", "user_defined_label_only"
+        ] = "canonical_label",
+        status_fetch_retry: int = STATUS_FETCH_RETRY_DEFAULT,
+        status_fetch_delay: int = STATUS_FETCH_DELAY_DEFAULT,
+    ) -> InventoryDevice[CustomSettings_com_nevion_spg9000_0_1_0]: ...
+
+    @overload
+    def get_device(
+        self,
+        label: Optional[str] = None,
+        device_id: Optional[str] = None,
+        address: Optional[str] = None,
         custom_settings_type: Optional[Literal["com.nevion.starfish_splicer-0.1.0"]] = None,
         config_only: bool = False,
         label_search_mode: Literal[
@@ -1638,6 +1713,21 @@ class InventoryGetDeviceMixin:
         status_fetch_retry: int = STATUS_FETCH_RETRY_DEFAULT,
         status_fetch_delay: int = STATUS_FETCH_DELAY_DEFAULT,
     ) -> InventoryDevice[CustomSettings_com_nevion_tag_mcm9000_0_1_0]: ...
+
+    @overload
+    def get_device(
+        self,
+        label: Optional[str] = None,
+        device_id: Optional[str] = None,
+        address: Optional[str] = None,
+        custom_settings_type: Optional[Literal["com.nevion.tag_mcs-0.1.0"]] = None,
+        config_only: bool = False,
+        label_search_mode: Literal[
+            "canonical_label", "factory_label_only", "user_defined_label_only"
+        ] = "canonical_label",
+        status_fetch_retry: int = STATUS_FETCH_RETRY_DEFAULT,
+        status_fetch_delay: int = STATUS_FETCH_DELAY_DEFAULT,
+    ) -> InventoryDevice[CustomSettings_com_nevion_tag_mcs_0_1_0]: ...
 
     @overload
     def get_device(
@@ -1870,6 +1960,36 @@ class InventoryGetDeviceMixin:
         label: Optional[str] = None,
         device_id: Optional[str] = None,
         address: Optional[str] = None,
+        custom_settings_type: Optional[Literal["com.nevion.txdarwin_dynamic-0.1.0"]] = None,
+        config_only: bool = False,
+        label_search_mode: Literal[
+            "canonical_label", "factory_label_only", "user_defined_label_only"
+        ] = "canonical_label",
+        status_fetch_retry: int = STATUS_FETCH_RETRY_DEFAULT,
+        status_fetch_delay: int = STATUS_FETCH_DELAY_DEFAULT,
+    ) -> InventoryDevice[CustomSettings_com_nevion_txdarwin_dynamic_0_1_0]: ...
+
+    @overload
+    def get_device(
+        self,
+        label: Optional[str] = None,
+        device_id: Optional[str] = None,
+        address: Optional[str] = None,
+        custom_settings_type: Optional[Literal["com.nevion.txdarwin_static-0.1.0"]] = None,
+        config_only: bool = False,
+        label_search_mode: Literal[
+            "canonical_label", "factory_label_only", "user_defined_label_only"
+        ] = "canonical_label",
+        status_fetch_retry: int = STATUS_FETCH_RETRY_DEFAULT,
+        status_fetch_delay: int = STATUS_FETCH_DELAY_DEFAULT,
+    ) -> InventoryDevice[CustomSettings_com_nevion_txdarwin_static_0_1_0]: ...
+
+    @overload
+    def get_device(
+        self,
+        label: Optional[str] = None,
+        device_id: Optional[str] = None,
+        address: Optional[str] = None,
         custom_settings_type: Optional[Literal["com.nevion.txedge-0.1.0"]] = None,
         config_only: bool = False,
         label_search_mode: Literal[
@@ -2058,6 +2178,21 @@ class InventoryGetDeviceMixin:
         status_fetch_retry: int = STATUS_FETCH_RETRY_DEFAULT,
         status_fetch_delay: int = STATUS_FETCH_DELAY_DEFAULT,
     ) -> InventoryDevice[CustomSettings_com_sony_SC1_1_0]: ...
+
+    @overload
+    def get_device(
+        self,
+        label: Optional[str] = None,
+        device_id: Optional[str] = None,
+        address: Optional[str] = None,
+        custom_settings_type: Optional[Literal["com.sony.XVS-G1-1.0"]] = None,
+        config_only: bool = False,
+        label_search_mode: Literal[
+            "canonical_label", "factory_label_only", "user_defined_label_only"
+        ] = "canonical_label",
+        status_fetch_retry: int = STATUS_FETCH_RETRY_DEFAULT,
+        status_fetch_delay: int = STATUS_FETCH_DELAY_DEFAULT,
+    ) -> InventoryDevice[CustomSettings_com_sony_XVS_G1_1_0]: ...
 
     @overload
     def get_device(
